Abstract plan data handler
